middleware: apply CORS before content type checker
A special case of xhr requests when the progress callback is used, Webkit and
Firefox will preform CORS checking on the response. If a request is CORS, the
response should contain Access-Control-Allow-Origin header set.

In our stack, multipart/form-data checker was applied before CORS middleware.
Hence, requests with incorrect Content-Type would be bounced by
multipart/form-data checker and never reach CORS middleware, where
Access-Control-Allow-Origin is appended. On the client side, it would look like
the reason for request to fail was not incorrect Content-Type, but a CORS check
failure.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>